pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["block_agents"]

[project]
name = "block-agent-backend"
version = "0.1.0"
description = "A modular, block-based system for creating and executing automated workflows"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Vittorio Rossi", email = "rossi.vittori@gmail.com" }
]

dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.5.0",
    "cohere>=4.0.0",
    "huggingface-hub>=0.16.0",
    "requests>=2.0.0",
    "python-multipart>=0.0.6",
    "python-docx>=0.8.11",
    "markdown>=3.4.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "pylint>=2.17.0",
    "ruff>=0.0.252"
]

[tool.uv]
exclude = []
index-url = "https://pypi.org/simple"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=block_agents"

[tool.ruff]
target-version = "py38"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "C4",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "ANN", # flake8-annotations
    "S",   # flake8-bandit
    "A",   # flake8-builtins
]
ignore = [
    "ANN101",  # Missing type annotation for `self` in method
    "ANN102",  # Missing type annotation for `cls` in classmethod
]
unfixable = [
    "F401",  # Don't automatically remove unused imports
    "F841",  # Don't automatically remove unused variables
]

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["S101", "ANN"]
"__init__.py" = ["F401"]